Extract HomePage class to its own file
