+++ lib/main.dart
@@ -2,13 +2,8 @@ import 'package:flutter/material.dart';
 
 import 'package:coffeecard/base/strings.dart';
 import 'package:coffeecard/base/style/theme.dart';
-import 'package:coffeecard/base/style/colors.dart';
-import 'package:coffeecard/base/style/text_styles.dart';
 
-import 'package:coffeecard/widgets/pages/tickets_page.dart';
-import 'package:coffeecard/widgets/pages/receipts_page.dart';
-import 'package:coffeecard/widgets/pages/stats_page.dart';
-import 'package:coffeecard/widgets/pages/settings_page.dart';
+import 'package:coffeecard/widgets/pages/home_page.dart';
 
 void main() {
   runApp(MyApp());
@@ -21,78 +16,6 @@ class MyApp extends StatelessWidget {
       title: Strings.appTitle,
       theme: analogTheme,
       home: HomePage(),
-    );
-  }
-}
-
-class HomePage extends StatefulWidget {
-  @override
-  _HomePageState createState() => _HomePageState();
-}
-
-class _HomePageState extends State<HomePage> {
-  int _currentPageIndex = 0;
-
-  List _pages = [
-    {
-      'appBarTitle': Strings.ticketsPageTitle,
-      'navBarTitle': Strings.ticketsNavTitle,
-      'icon': Icons.style,
-      'body': TicketsPage()
-    },
-    {
-      'appBarTitle': Strings.receiptsPageTitle,
-      'navBarTitle': Strings.receiptsNavTitle,
-      'icon': Icons.receipt,
-      'body': ReceiptsPage()
-    },
-    {
-      'appBarTitle': Strings.statsPageTitle,
-      'navBarTitle': Strings.statsNavTitle,
-      'icon': Icons.trending_up,
-      'body': StatsPage()
-    },
-    {
-      'appBarTitle': Strings.settingsPageTitle,
-      'navBarTitle': Strings.settingsNavTitle,
-      'icon': Icons.settings,
-      'body': SettingsPage()
-    },
-  ];
-
-  get _currentPage => _pages[_currentPageIndex];
-
-  void _onBottomNavTapped(int index) {
-    setState(() {
-      _currentPageIndex = index;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: AppColor.background,
-      appBar: AppBar(
-        title: Text(
-          _currentPage['appBarTitle'],
-          style: AppTextStyle.pageTitle
-        ),
-      ),
-      body: _currentPage['body'],
-      bottomNavigationBar: BottomNavigationBar(
-        currentIndex: _currentPageIndex,
-        onTap: _onBottomNavTapped,
-        type: BottomNavigationBarType.fixed,
-        backgroundColor: AppColor.primary,
-        selectedItemColor: AppColor.white,
-        unselectedItemColor: AppColor.white.withOpacity(0.5),
-        selectedFontSize: 12,
-        items: List.generate(_pages.length, (index) => BottomNavigationBarItem(
-            icon: Icon(_pages[index]['icon']),
-            title: Text(_pages[index]['navBarTitle'], style: AppTextStyle.medium)
-          )
-        ),
-      )
     );
   }
 }
